Change X-axis scaling decorator for teal.modules.clinical::tm_g_forest_rsp.

A function to create a UI component for selecting a transform function for the forest plot x axis.

Usage

forestplot_x_decorator()

Value

Returns a modified plot object with the transformation applied.

Details

The module creates a UI with a radio control for selecting the transform function. The selected transformation function is applied to the forest plot to update the plot's axis and annotations accordingly.

Examples

data <- teal_data()
#> Error in teal_data(): could not find function "teal_data"
data <- within(data, {
  ADSL <- tmc_ex_adsl
  ADRS <- tmc_ex_adrs %>%
    dplyr::mutate(AVALC = tern::d_onco_rsp_label(AVALC) %>%
      dplyr::with_label("Character Result/Finding")) %>%
    dplyr::filter(PARAMCD != "OVRINV" | AVISIT == "FOLLOW UP")
})
#> Error in UseMethod("within"): no applicable method for 'within' applied to an object of class "function"
join_keys(data) <- default_cdisc_join_keys[names(data)]
#> Error: object 'default_cdisc_join_keys' not found

ADSL <- data[["ADSL"]]
#> Error in data[["ADSL"]]: object of type 'closure' is not subsettable
ADRS <- data[["ADRS"]]
#> Error in data[["ADRS"]]: object of type 'closure' is not subsettable

arm_ref_comp <- list(
  ARM = list(
    ref = "B: Placebo",
    comp = c("A: Drug X", "C: Combination")
  ),
  ARMCD = list(
    ref = "ARM B",
    comp = c("ARM A", "ARM C")
  )
)

app <- teal::init(
  data = data,
  modules = teal::modules(
    teal.modules.clinical::tm_g_forest_rsp(
      label = "Forest Response",
      dataname = "ADRS",
      arm_var = choices_selected(
        variable_choices(ADSL, c("ARM", "ARMCD")),
        "ARMCD"
      ),
      arm_ref_comp = arm_ref_comp,
      paramcd = choices_selected(
        value_choices(ADRS, "PARAMCD", "PARAM"),
        "INVET"
      ),
      subgroup_var = choices_selected(
        variable_choices(ADSL, names(ADSL)),
        c("BMRKR2", "SEX")
      ),
      strata_var = choices_selected(
        variable_choices(ADSL, c("STRATA1", "STRATA2")),
        "STRATA2"
      ),
      plot_height = c(600L, 200L, 2000L),
      decorators = list(
        plot = forestplot_x_decorator()
      ),
      default_responses = list(
        BESRSPI = list(
          rsp = c("Stable Disease (SD)", "Not Evaluable (NE)"),
          levels = c(
            "Complete Response (CR)", "Partial Response (PR)", "Stable Disease (SD)",
            "Progressive Disease (PD)", "Not Evaluable (NE)"
          )
        ),
        INVET = list(
          rsp = c("Complete Response (CR)", "Partial Response (PR)"),
          levels = c(
            "Complete Response (CR)", "Not Evaluable (NE)", "Partial Response (PR)",
            "Progressive Disease (PD)", "Stable Disease (SD)"
          )
        ),
        OVRINV = list(
          rsp = c("Progressive Disease (PD)", "Stable Disease (SD)"),
          levels = c("Progressive Disease (PD)", "Stable Disease (SD)", "Not Evaluable (NE)")
        )
      )
    )
  )
)
#> Error in teal::init(data = data, modules = teal::modules(teal.modules.clinical::tm_g_forest_rsp(label = "Forest Response",     dataname = "ADRS", arm_var = choices_selected(variable_choices(ADSL,         c("ARM", "ARMCD")), "ARMCD"), arm_ref_comp = arm_ref_comp,     paramcd = choices_selected(value_choices(ADRS, "PARAMCD",         "PARAM"), "INVET"), subgroup_var = choices_selected(variable_choices(ADSL,         names(ADSL)), c("BMRKR2", "SEX")), strata_var = choices_selected(variable_choices(ADSL,         c("STRATA1", "STRATA2")), "STRATA2"), plot_height = c(600L,         200L, 2000L), decorators = list(plot = forestplot_x_decorator()),     default_responses = list(BESRSPI = list(rsp = c("Stable Disease (SD)",         "Not Evaluable (NE)"), levels = c("Complete Response (CR)",         "Partial Response (PR)", "Stable Disease (SD)", "Progressive Disease (PD)",         "Not Evaluable (NE)")), INVET = list(rsp = c("Complete Response (CR)",         "Partial Response (PR)"), levels = c("Complete Response (CR)",         "Not Evaluable (NE)", "Partial Response (PR)", "Progressive Disease (PD)",         "Stable Disease (SD)")), OVRINV = list(rsp = c("Progressive Disease (PD)",         "Stable Disease (SD)"), levels = c("Progressive Disease (PD)",         "Stable Disease (SD)", "Not Evaluable (NE)")))))): Assertion on 'data' failed: Must inherit from class 'teal_data'/'teal_data_module', but has class 'function'.
if (interactive()) {
  shinyApp(app$ui, app$server)
}